Implemented comparing of encrypted hashes

+++ index.js
@@ -21,20 +21,21 @@ app.post("/" + config.hook_path, (req, res) => {
     console.log(req.rawHeaders)
     console.log(req.rawHeaders[19])
 
-    hmac = crypto.createHmac('sha1', config.password)
-    hmac.update(req.body.toString())
-    var check = hmac.digest('hex')
+    var check = `sha1=${crypto.createHmac('sha1', config.hook_secret).update(JSON.stringify(req.body), 'utf-8').digest('hex')}`
+    console.log(check);
+    if (check === req.rawHeaders[19]) {
+        fs.removeSync(config.clone_directory)
+        git.Clone(req.body.repository.clone_url, config.clone_directory).then(function (repository) {
+            // Work with the repository object here.
+        });
+        res.status(200).end() // Responding is important
+    } else {
+        res.status(403).end()
+    }
 
-    console.log("sha1=" + check);
-    console.log(check === req.rawHeaders[19])
-    
-    fs.removeSync(config.clone_directory)
-    git.Clone(req.body.repository.clone_url, config.clone_directory).then(function (repository) {
-        // Work with the repository object here.
-    });
-    res.status(200).end() // Responding is important
+
 })
 
 // Start express on the defined port
 console.log("Listening on:")
-app.listen(PORT, () => console.log(`http://${ip.address()}:${PORT}`+"/" + config.hook_path ))
+app.listen(PORT, () => console.log(`http://${ip.address()}:${PORT}` + "/" + config.hook_path))
